CLI/requirements.txt
psycopg2-binary==2.9.10
google-api-python-client==2.154.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1

